relatorios-api/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
geoalchemy2==0.14.2

# PDF Generation
weasyprint==60.1
reportlab==4.0.8

# AWS S3
boto3==1.34.10

# Auth
python-jose[cryptography]==3.3.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utils
jinja2==3.1.2
matplotlib==3.8.2
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
